Refactors, add mutable DoubleLinkedList :(
