+++ src/main/scala/common/Node.scala
@@ -1,0 +1,5 @@
+package common
+
+trait Node[T] {
+  def value: T
+}
